Support for encrypted keys
Previously the system did not properly support encrypted keys, so now you will be prompted for the passphrase of your key as appropriate during the process. Unfortunately this means if there are multiple pages that you will be prompted twice. We hope to work around this problem in the future.